backend_fastapi/pyproject.toml
[project]
name = "healthcart-backend"
version = "1.0.0"
description = "HealthCart Telemedicine Platform Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.11.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "httpx>=0.28.0",
    "razorpay>=1.4.2",
    "weasyprint>=63.0",
    "jinja2>=3.1.4",
    "firebase-admin>=6.6.0",
    "paho-mqtt>=2.1.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
